sine_function/requirements.txt
cognite-extractor-utils>=2.2.0
cognite-logger>=0.5
numpy>=1.21.6